+++ scripts/sql/tables/publish_count.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `publish_count` (
-    `user_id` INT(11) NOT NULL,
-    `count` INT(11) NOT NULL,
+    `user_id` BIGINT NOT NULL,
+    `count` BIGINT NOT NULL,
     PRIMARY KEY (`user_id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8mb4 COLLATE=utf8mb4_unicode_ci;
